app: saf: Decouple SAF from Zephyr SPI driver API
Allow no SPI initialization via SPI driver sequences
Move SPI flash initialization in separate file
Add app saf module CMakelist to allow library in the future
